makefile: Install from distribution files
Installing with `cp` and parallel copies seems to cause almost
systematic disk corruption on DM32 with macOS Sequoia. After
installation, the disk shows 0B free and won't mount without `fsck`.
Even `fsck` is not sufficient to make the disk writable again.
The `config` directory typically changes to a file.

Apparently, extracting the files sequentially with `tar` does not
suffer the same problem. It also ensures the structure of the
installation file is correct.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>