ui: Refactor methods to insert objects in the text editor
There were a mix of `edit` and `insert`, and two distinct ways to copy
an object in the editor, one of them wastelessly createing a text
object as an intermediary.

This resulted in a number of rather silly bugs that were first exposed
while testing the `Custom` menu (where we for the first time can put
arbitrary objects in a menu), but upon reflection can also be
triggered relatively easily in the interactive stack when using `Echo`
when not at the end of the command line.

Fixes: #1256

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>