equations: Add `assignment` objects
Assignment objects have the form `Name=Value`, `Name◀Value`,
`Name←Value`, `Value▶Name` or `Value→Name`. In expressions, the form
`Name=Value` is not accepted, it is interpreted as the equality
operator.

An assignment evaluates `Value`, stores it in `Name` and returns the
tagged `:Name:Value` object on the stack.

This is intended primarily to setup library equations in a readable
way, for example:

```
m=1.5_g
'ROOT(ⒺRelativity Mass Energy;E;1_MJ)' EVAL
```

A side effect is a change in the random number generator because of
the change in the type ID for `decimal`, which is used by `RandomSeed`

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>