ui: Add ability to customize the header
The `Header` command updates a special variable also called `Header`.

When that variable is present, it must evaluate to something that can
render graphically, either directly a graphic object or a (possibly
multi-line) text.

When a header is provided, the normal content of the header,
i.e. date, time and name of the state file, is no longer shown.
However, annunciators and battery status are still overimposed.

It is the responsibility of the programmer to ensure that the header
program does not draw important data at these locations, and also to
make sure that the header program is "well behaved", i.e. does not
leave things on stack. If the header program generates an error, then
that error may get in the way of normal calculator operations.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>